Add mongoose server prototype
This was a huge pain in the ass and I hated every second of it. cmake did NOT want to compile a single C file, without trying to compile everything else in the wrong mode. Compiling mongoose as C++ was way more straight forward, after figuring out how to disable the warnings.

MSVC does not actually respect overwriting the warning flag, despite what it says.